Photography/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
